+++ helpers/settings.py
@@ -1,11 +1,3 @@
-"""
-DRAFT
-
-Eventually we'll pull things from here. It is meant to easier change settings
-
-Right now everything is still in it's place
-"""
-
 # Which extensions to load on startup
 extensions = [
     'cogs.giveaway',
@@ -21,19 +13,30 @@ extensions = [
 # Status message for bot
 now_playing = "!help"
 
-# Identities
+# ID's
 bot_id = "180842549873737728"
-owner_id = "132577770046750720"
+dev_id = "132577770046750720"
+
+server_owner = "Idiotech"
+owner_id = "176291669254209539"
+
+# Used to get Idiotech's social media stuff -
+twitch = "idiotechgaming"
+twitter_handle = "idiotechgaming"
+facebook = "idiotechgaming"
+yt_channel = "UC0YagOInbZxj10gaWwb1Nag"
+reddit = "idiotechgaming"
+
 
 channels = {
     "admin": "176304607172100097",
     "mods": "188172983699767296",
     "news": "179963038781276160",
-    "general": "176293292865093632",
+    "general": "165120765283270656",
     "rules": "179965419728273408",
     "gaming": "179945195280924673",
     "tabletop": "188081332834926592",
-    "technology": "188085243478147074",
+    "technology": "188083195470807040",
     "programming": "188083195470807040",
     "giveaways": "188087131070005248",
 }