+++ cogs/versioning/changelog.md
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+1.0.3
+=====
+* Added !e3 which returns an E3 timetable
+* Added [!steam][deals] - returns link to webpage will list of sales on
+* Fixed giveaway topic not updating after a giveaway was closed (bug)
+* A few more "under the hood" changes
+
+1.0.2.1
+=======
+* Made a few visual changes (check github if interested)
+* A few "under the hood" changes
+* Added [!steam][bestsellers] which returns top 5 best selling games on Steam
+* Moved what was [!steam] to [!steam][status]
+
+
 1.0.2
 =====